BFT-464: Move to parts.rs
